Replace near/far by depthNear/depthFar. Fix compilation on Windows.
Reason:
Generated C++ code from IDL cannot compile on Windows since near
and far are reserved keywords MS compiler.

It might be possible to undef it in C++ headers but this seems more
appropriate since:

MediaTrackSettings and other dictionaries are generic, meaning that
they apply to all stream tracks and it is good to explicitelly mark
that this applies only to depth. On the other side focalLength,
principal point and other intrinsics that we plan to add apply to
all cameras so there is no need to prefix them with depth.

This way, it is also in sync with depthNear and depthFar in WebVR
API's depthNear/depthFar[1].

[1]
https://w3c.github.io/webvr/#dom-vrdisplay-depthnear