Depth value - distance conversion. Grayscale and 16-bit depth removed.
Current approach is to push the original data (captured by camera)
to the WebGL textures without any conversion to a "capturable subrange"
as it would introduce performance penalty and it would not be available
on all of the devices so removing grayscale rendering.

Part of grayscale computation is also related to depth - distance:
kept and simplified.

Using 16-bit as the depth map format is not needed - normalized value
in range (0-1] is better and could handle the float point cloud to
float depth map conversion for Tango.

Fix #142